Adds semi-functional command to become the fair_child.
There is a problem with the asynchronous functions that is
  causing the response to the fair child and setting the
  fair child to occur at the same time.
The best method to resolve this would probably be to stop
on_message from responding if the message is a command.
I'm not aware of an easy "discord.py" way to do this, but
regex of the first word against recognized previxes might
work.